Made CaseGroupInfo and CasePageInfo top-level classes.
Created a new package to hold models used only in the service layer,
and put the two types that used to be inner classes of CaseListService
in there instead.
