+++ src/main/java/gov/usds/case_issues/services/PageTranslationService.java
@@ -1,7 +1,7 @@
 package gov.usds.case_issues.services;
 
-import gov.usds.case_issues.services.CaseListService.CaseGroupInfo;
-import gov.usds.case_issues.services.CaseListService.CasePageInfo;
+import gov.usds.case_issues.services.model.CaseGroupInfo;
+import gov.usds.case_issues.services.model.CasePageInfo;
 import gov.usds.case_issues.validators.TagFragment;
 
 public interface PageTranslationService {
